Xeno-canto snapshot samples extraction
This Jupyter notebook shows how to extract detected events from the Xeno-canto snapshot from Birdset. The code is structured to subset the database to a target species and generate a dataframe containing the audio paths as well as the start time of the peaks detected (automatically detected and not verified). Moreover, it provide the option to generate a list of confusion species based on a geographical filter